requirements.txt
python-binance==1.0.16
PyYAML
Flask
requests
pandas
ta
python-dotenv
